[C] A Lexical Analyzer for Light_C
The program emulates a lexical analyzer, in C programming language.
At first, it scans the first letter of the user-input string, and it
calls the appropriate function for the job (variables, comment, number,
strings).
Each function, built by the template of its DFA
(http://en.wikipedia.org/wiki/DFSA),
then checks the string, and outputs the validity and charlength of the
input.
If there is a text file named 'input.txt' on its directory, the program
will also
scan the file's contents as lines, outputing the information (type and
size) to a
new file named 'output.txt'.
f.e input: 6.478932 \n //com// \n "etc." | output: NUMBER 8 \n COMMENT 3
\n STRING 4

Compiled with DEV-C++ {mingw32-gcc-3.4.2} on a Windows 7 x64 system.
For TEI of Athens and Mrs. Georgouli, a project for Compilers class.
©Anastasios Georgopoulos, Athens 2013