An .l input file for FLEX (Light_C)
This .l file is built to be an input file for the FLEX (Fast LEXical
analyzer) tool.
In short, the result we want is to instruct FLEX to perform
pattern-matching on text,
giving it the REGular EXpressions to do so. The result is an outline of
the source code,
with each piece identified correctly. The syntax style is C-like,
customed to FLEX's requirements.
Now including every symbol or reserved word there is in Light-C! (REV2.0
9/12/13)
More about the proper format of a FLEX input file can be found on
http://flex.sourceforge.net/manual/

Version 1.0 was compiled with the last version of FLEX on the Ubuntu
repository as of 1st Dec 2013,
and the result C file was complied by gcc tool, on a Ubuntu 12.03 x64
system.
Version 2.0 was compiled with the last version of FLEX on the Ubuntu
repository as of 9th Dec 2013,
and the result C file was complied by gcc tool, on a Ubuntu 12.10 x64
system.

For TEI of Athens and Mrs. Georgouli, a project for Compilers class.
?©Anastasios Georgopoulos, Athens 2013